+++ test/iot-ts-aggregates.spec.ts
@@ -1,11 +1,12 @@
 import { MindSphereSdk, TimeSeriesAggregateModels } from "../src/api/sdk";
 import { checkAssetId, decrypt, loadAuth } from "../src/api/utils";
 import { setupStructure } from "./test-agent-setup-utils";
+import { getPasskeyForUnitTest } from "./test-utils";
 
 describe("[SDK] IotTsAggregateClient", () => {
     const auth = loadAuth();
 
-    const sdk = new MindSphereSdk({ ...auth, basicAuth: decrypt(auth, "passkey.4.unit.test") });
+    const sdk = new MindSphereSdk({ ...auth, basicAuth: decrypt(auth, getPasskeyForUnitTest()) });
 
     let assetid = "";